refactor: lift internal/app to public app package
Downstream binaries (cyoda-go-cassandra's full binary in Plan 4c)
need to assemble a cyoda-go App with their own set of blank-imported
storage plugins. That requires app.New, app.DefaultConfig,
app.LoadEnvFiles, etc. to be reachable from outside the cyoda-go
module — which internal/ prevents.

Mechanical move: internal/app → app. Thirteen in-tree consumers had
their import paths rewritten. No API changes. All tests green.

Note: App's getter methods still return types from internal/*
packages (contract.AuditService, *internalgrpc.Server, etc.). Go
permits calling such methods from external callers; only naming the
return types in a variable is forbidden. The full-binary pattern
(a.GRPCServer().Serve(lis)) works without exposing those internals.
