+++ internal/domain/messaging/handler_test.go
@@ -8,7 +8,7 @@ import (
 	"strings"
 	"testing"
 
-	"github.com/cyoda-platform/cyoda-go/internal/app"
+	"github.com/cyoda-platform/cyoda-go/app"
 
 	_ "github.com/cyoda-platform/cyoda-go/plugins/memory"
 )
