+++ internal/iam/mock/mock_test.go
@@ -8,7 +8,7 @@ import (
 
 	spi "github.com/cyoda-platform/cyoda-go-spi"
 	"github.com/cyoda-platform/cyoda-go/internal/api/middleware"
-	"github.com/cyoda-platform/cyoda-go/internal/app"
+	"github.com/cyoda-platform/cyoda-go/app"
 
 	mockiam "github.com/cyoda-platform/cyoda-go/internal/iam/mock"
 	_ "github.com/cyoda-platform/cyoda-go/plugins/memory"
